Module tor_hscrypto::pk

Key type wrappers of various kinds used in onion services.

(We define wrappers here as a safety net against confusing one kind of key for another: without a system like this, it can get pretty hard making sure that each key is used only in the right way.)

Structs

HsBlindId

A blinded onion service identity, represented in a compact format. (KP_hs_blind_id)

HsBlindIdKey

The “blinded” identity of a v3 onion service. (KP_hs_blind_id)

HsBlindIdKeypair

The private counterpart of a [`HsBlindIdKeyKey’].

HsClientDescEncKey

Second type of client authorization key, used for onion descriptor decryption. (KP_hsc_desc_enc)

HsClientDescEncSecretKey

The private counterpart of a [`HsClientDescEncKeyKey’].

HsClientIntroAuthKey

First type of client authorization key, used for the introduction protocol. (KP_hsc_intro_auth)

HsClientIntroAuthKeypair

The private counterpart of a [`HsClientIntroAuthKeyKey’].

HsId

The identity of a v3 onion service. (KP_hs_id)

HsIdKey

The identity of a v3 onion service, expanded into a public key. (KP_hs_id)

HsIdKeypair

The private counterpart of a [`HsIdKeyKey’].

HsIntroPtSessionIdKey

A key used to identify and authenticate an onion service at a single introduction point. (KP_hs_ipt_sid)

HsIntroPtSessionIdKeypair

The private counterpart of a [`HsIntroPtSessionIdKeyKey’].

HsSvcDescEncKey

Server key, used for diffie hellman during onion descriptor decryption. (KP_hss_desc_enc)

HsSvcDescEncKeypair

An ephemeral x25519 keypair, generated by an onion service and used to for onion service encryption.

HsSvcDescEncSecretKey

The private counterpart of a [`HsSvcDescEncKeyKey’].

HsSvcNtorKey

A key used in the HsNtor handshake between the client and the onion service. (KP_hss_ntor)

HsSvcNtorSecretKey

The private counterpart of a [`HsSvcNtorKeyKey’].

Enums

HsIdParseError

Error that can occur parsing an HsId from a v3 .onion domain name

Constants

HSID_ONION_SUFFIX

The fixed string .onion