Struct tor_hscrypto::pk::HsClientDescEncKey

pub struct HsClientDescEncKey(_);

Second type of client authorization key, used for onion descriptor decryption. (KP_hsc_desc_enc)

Any client who knows the secret key corresponding to this key can decrypt the inner layer of the onion service descriptor.